+++ public_html/index.php
@@ -43,7 +43,7 @@ $expected = array(
     // Views
     'products', 'order', 'view', 'detail', 'printorder',
     'orderhist', 'packinglist',
-    'couponlog',
+    'couponlog', 'category',
     'cart', 'pidetail', 'viewcart',
 );
 $action = 'products';    // default view
@@ -353,12 +353,20 @@ case 'checkoutcart':
 
 case 'products':
 default:
-    USES_shop_functions();
-    SHOP_setUrl();
     $cat_id = isset($_REQUEST['category']) ? (int)$_REQUEST['category'] : 0;
-    $content .= \Shop\ProductList($cat_id);
-    $menu_opt = $LANG_SHOP['products'];
-    $page_title = $LANG_SHOP['main_title'];
+    if (
+        ($_SHOP_CONF['hp_layout'] & SHOP_HP_CAT) == SHOP_HP_CAT &&
+        $cat_id == 0 &&
+        (!isset($_GET['query']) || isset($_GET['clearsearch']))
+    ) {
+        $content .= SHOP_homepage_category();
+    } else {
+        USES_shop_functions();
+        SHOP_setUrl();
+        $content .= \Shop\ProductList($cat_id);
+        $menu_opt = $LANG_SHOP['products'];
+        $page_title = $LANG_SHOP['main_title'];
+    }
     break;
 
 case 'none':
@@ -376,5 +384,56 @@ $display .= $T->parse('', 'title');
 $display .= $content;
 $display .= \Shop\Menu::siteFooter();
 echo $display;
+exit;
+
+
+/**
+ * Display the shop home page as a collection of category tiles.
+ *
+ * @return  string  HTML for category homepage
+ */
+function SHOP_homepage_category()
+{
+    global $_SHOP_CONF;
+
+    $display = '';
+    $cat_sql = '';
+
+    $RootCat = \Shop\Category::getRoot();
+    $Cats = \Shop\Category::getTree();
+    $T = new \Template(SHOP_PI_PATH . '/templates');
+    $T = SHOP_getTemplate(array(
+        'wrapper'   => 'list/' . $_SHOP_CONF['list_tpl_ver'] . '/wrapper',
+        'start'   => 'product_list_start',
+        'end'     => 'product_list_end',
+    ) );
+    $T->set_var('pi_url', SHOP_URL);
+
+    $T->set_block('wrapper', 'ProductItems', 'PI');
+    foreach ($Cats as $Cat) {
+        // If this is the root category, and root shouldn't be included,
+        // then skip it.
+        if (
+            $Cat->cat_id == $RootCat->cat_id &&
+            ($_SHOP_CONF['hp_layout'] & SHOP_HP_CATHOME) != SHOP_HP_CATHOME
+        ) {
+            continue;
+        }
+        $T->set_var(array(
+            'item_id'       => $Cat->cat_id,
+            'short_description' => htmlspecialchars($Cat->cat_name),
+            'img_cell_width' => ($_SHOP_CONF['max_thumb_size'] + 20),
+            'item_url'      => SHOP_URL . '/index.php?category='. $Cat->cat_id,
+            'small_pic'     => $Cat->ImageUrl(),
+            'tpl_ver'       => $_SHOP_CONF['list_tpl_ver'],
+        ) );
+        $T->parse('PI', 'ProductItems', true);
+        //var_dump($T);die;
+    }
+    $display .= $T->parse('', 'start');
+    $display .= $T->parse('', 'wrapper');
+    $display .= $T->parse('', 'end');
+    return $display;
+}
 
 ?>